Blog Comment 관리 프로그램 Lv2 과제 완성

+++ src/main/java/com/sparta/blog/service/CommentService.java
@@ -4,6 +4,7 @@ import com.sparta.blog.dto.CommentRequestDto;
 import com.sparta.blog.dto.CommentResponseDto;
 import com.sparta.blog.entity.Comment;
 import com.sparta.blog.repository.CommentRepository;
+import com.sparta.blog.security.UserDetailsImpl;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -19,9 +20,9 @@ public class CommentService {
         this.commentRepository = commentRepository;
     }
 
-    public CommentResponseDto createComment(CommentRequestDto requestDto) {
+    public CommentResponseDto createComment(CommentRequestDto requestDto, UserDetailsImpl userDetails) {
         // RequestDto -> Entity
-        Comment comment = new Comment(requestDto);
+        Comment comment = new Comment(requestDto, userDetails);
         // DB 저장
         Comment saveComment = commentRepository.save(comment);
         // Entity -> ResponseDto
@@ -44,22 +45,30 @@ public class CommentService {
     }
 
     @Transactional
-    public CommentRequestDto updateComment(Long id, CommentRequestDto requestDto) {
+    public CommentRequestDto updateComment(Long id, CommentRequestDto requestDto, UserDetailsImpl userDetails) {
         // 해당 메모가 DB에 존재하는지 확인
         Comment comment = findComment(id);
-        // comment 내용 수정
-        comment.update(requestDto);
+        if (comment.getUsername().equals(userDetails.getUsername())) {
+            // comment 내용 수정
+            comment.update(requestDto);
+        } else {
+            throw new IllegalArgumentException("해당 게시글의 작성자가 아닙니다.");
+        }
 
         return requestDto;
     }
 
-    public String deleteComment(Long id) {
+    public String deleteComment(Long id, UserDetailsImpl userDetails) {
         // 해당 메모가 DB에 존재하는지 확인
         Comment comment = findComment(id);
-        // comment 삭제
-        commentRepository.delete(comment);
-
-        return "게시글을 삭제하는 데 성공하였습니다.";
+        if (comment.getUsername().equals(userDetails.getUsername())) {
+            // comment 삭제
+            commentRepository.delete(comment);
+            return "게시글을 삭제하는 데 성공하였습니다.";
+        } else {
+//            throw new IllegalArgumentException("해당 게시글의 작성자가 아닙니다.");
+            return "redirect:/blog";
+        }
     }
 
     private Comment findComment(Long id) {
